Post-deployment tweaks and bug fixes (#142)
* Fix bug with failed job aggregation
* cancel separately

+++ trailblazer/store/api.py
@@ -217,18 +217,17 @@ class BaseHandler:
             self.commit()
         return old_analyses
 
-    def delete_analysis(self, analysis_id: int, force: bool = False, ssh: bool = False) -> None:
+    def delete_analysis(self, analysis_id: int, force: bool = False) -> None:
         """Delete the analysis output."""
         analysis_obj = self.analysis(analysis_id=analysis_id)
         if not analysis_obj:
             raise TrailblazerError("Analysis not found")
-        self.update_run_status(analysis_id=analysis_id, ssh=ssh)
+
         if not force and analysis_obj.status in ONGOING_STATUSES:
             raise TrailblazerError(
                 f"Analysis for {analysis_obj.family} is currently running! Use --force flag to delete anyway."
             )
         LOG.info(f"Deleting analysis {analysis_id} for case {analysis_obj.family}")
-        self.cancel_analysis(analysis_id=analysis_id, ssh=ssh)
         analysis_obj.delete()
         self.commit()
 
@@ -269,7 +268,7 @@ class BaseHandler:
     @staticmethod
     def get_time_elapsed_in_min(elapsed_string: Optional[str]) -> Optional[int]:
         """Parse SLURM elapsed time string into minutes"""
-        if elapsed_string:
+        if elapsed_string and isinstance(elapsed_string, str):
             days = 0
             if "-" in elapsed_string:
                 days = int(elapsed_string.split("-")[0])
@@ -282,6 +281,7 @@ class BaseHandler:
                 + days * 60
             )
             return elapsed_minutes
+        return 0
 
     @staticmethod
     def parse_squeue_to_df(squeue_response: Any, ssh: bool) -> pd.DataFrame:
@@ -313,6 +313,8 @@ class BaseHandler:
 
     def update_jobs(self, analysis_obj: models.Analysis, jobs_dataframe: pd.DataFrame) -> None:
         """Parses job dataframe and creates job objects"""
+        if len(jobs_dataframe) == 0:
+            return
         for job_obj in analysis_obj.failed_jobs:
             job_obj.delete()
         self.commit()
@@ -335,11 +337,19 @@ class BaseHandler:
         """Iterate over all analysis with ongoing status and query SLURM for current progress"""
         ongoing_analyses = self.analyses(temp=True)
         for analysis_obj in ongoing_analyses:
-            self.update_run_status(analysis_id=analysis_obj.id, ssh=ssh)
+            try:
+                self.update_run_status(analysis_id=analysis_obj.id, ssh=ssh)
+            except Exception as e:
+                LOG.error(
+                    f"Failed to update {analysis_obj.family} - {analysis_obj.id}: {e.__class__.__name__}"
+                )
 
     def update_run_status(self, analysis_id: int, ssh: bool = False) -> None:
         """Query slurm for entries related to given analysis, and update the Trailblazer database"""
         analysis_obj = self.analysis(analysis_id)
+        if not analysis_obj:
+            LOG.warning(f"Analysis {analysis_id} not found!")
+            return
         try:
             jobs_dataframe = self.parse_squeue_to_df(
                 squeue_response=self.query_slurm(
@@ -353,7 +363,7 @@ class BaseHandler:
                 jobs_dataframe.status.value_counts() / len(jobs_dataframe), 2
             )
 
-            LOG.info("Status in SLURM")
+            LOG.info(f"Status in SLURM: {analysis_obj.family} - {analysis_id}")
             LOG.info(jobs_dataframe)
             analysis_obj.progress = float(status_distribution.get("COMPLETED", 0.0))
             if status_distribution.get("FAILED") or status_distribution.get("TIMEOUT"):
@@ -361,13 +371,13 @@ class BaseHandler:
                     analysis_obj.status = "error"
                     analysis_obj.comment = (
                         f"WARNING! Analysis still running with failed steps: "
-                        f"{ ', '.join(list(jobs_dataframe[jobs_dataframe.status == 'FAILED']))}"
+                        f"{ ', '.join(list(jobs_dataframe[jobs_dataframe.status == 'FAILED']['step']))}"
                     )
                 else:
                     analysis_obj.status = "failed"
                     analysis_obj.comment = (
                         f"Failed steps: "
-                        f"{', '.join(list(jobs_dataframe[jobs_dataframe.status == 'FAILED']))}"
+                        f"{', '.join(list(jobs_dataframe[jobs_dataframe.status == 'FAILED']['step']))}"
                     )
 
             elif status_distribution.get("COMPLETED") == 1:
@@ -412,9 +422,9 @@ class BaseHandler:
 
             analysis_obj.logged_at = dt.datetime.now()
         except Exception as e:
-            LOG.error(f"Error logging case - {e}")
+            LOG.error(f"Error logging case - {e.__class__.__name__}")
             analysis_obj.status = "error"
-            analysis_obj.comment = f"Error logging case - {e}"
+            analysis_obj.comment = f"Error logging case - {e.__class__.__name__}"
             self.commit()
 
     @staticmethod
@@ -447,7 +457,7 @@ class BaseHandler:
         analysis_obj.status = "canceled"
         analysis_obj.comment = (
             f"Analysis cancelled manually by user:"
-            f" {(self.user(email).name if self.user(email) else 'Unknown')}!"
+            f" {(self.user(email).name if self.user(email) else (email or 'Unknown'))}!"
         )
         self.commit()
